+++ packages/docusaurus-theme-live-codeblock/src/types.d.ts
@@ -6,3 +6,13 @@
  */
 
 /// <reference types="@docusaurus/theme-classic" />
+
+declare module '@theme-init/CodeBlock' {
+  import type CodeBlock, {Props as BaseProps} from '@theme/CodeBlock';
+
+  export interface Props extends BaseProps {
+    live?: boolean;
+  }
+  const CodeBlockComp: typeof CodeBlock;
+  export default CodeBlockComp;
+}
